digital_bath/requirements.txt
Flask
paho-mqtt
numpy